poison_methods/GASLITE/requirements.txt
numpy==1.24.3
pandas==2.0.3
PyYAML==6.0.1
torch==2.0.1
transformers==4.44.2

accelerate==0.28.0
datasets==2.20.0
beir==2.0.0
huggingface_hub==0.24.6

hydra-core==1.3.2
omegaconf==2.3.0
tqdm
typer==0.12.5

kmedoids==0.5.1
scikit-learn==1.3.2
sentence_transformers==2.7.0
scikit-learn-extra
# pip install pytrec-eval-terrier  # for windows

